refactor(file-upload): 순수 함수를 utils로 이동하여 프로젝트 구조 규칙 준수
- getFileExtension과 validateFileExtension을 utils/mediaType.ts로 이동
- useFileUpload.ts에서 순수 함수 제거 및 utils에서 import
- utils/mediaUpload.ts에서 getFileExtension을 utils/mediaType.ts에서 import
- inferMimeType 함수에 JSDoc 주석 추가
- 프로젝트 구조 규칙 준수: utils/는 Pure Functions, components/는 Business Logic
